# src/domain/user/User.js
const { attributes } = require('structure');

const User = attributes({
    id: String,
    name: {
        type: String,
        required: true
    },
    age: Number
})(class User {
    isLegal() {
        return this.age >= User.MIN_LEGAL_AGE;
    }
});

User.MIN_LEGAL_AGE = 21;

module.exports = User;