# src/interfaces/http/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Node API Microservice Reference Architecture",
      "version": "v1"
    },
    "servers": [
      {
        "description": "Local server",
        "url": "/api"
      }
    ],
    "paths": {
      "/users": {
        "get": {
          "operationId": "listUsers",
          "tags": [ "Users" ],
          "responses": {
            "200": {
              "description": "List of all users",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "operationId": "createUser",
          "tags": [ "Users" ],
          "requestBody": {
            "description": "User data",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewUser"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "400": {
              "description": "User not created because of validation error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "operationId": "showUser",
          "tags": [ "Users" ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id of user to show",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Return user with given id",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "User not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/NotFoundError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "operationId": "updateUser",
          "tags": [ "Users" ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id of user to update",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "requestBody": {
            "description": "User new data",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewUser"
                }
              }
            }
          },
          "responses": {
            "202": {
              "description": "User updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "User not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/NotFoundError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "operationId": "deleteUser",
          "tags": [ "Users" ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id of user to delete",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "202": {
              "description": "User deleted successfully"
            },
            "404": {
              "description": "User not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "allOf": [
            { "$ref": "#/components/schemas/NewUser" },
            {
              "required": [ "id" ],
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          ]
        },
        "NewUser": {
          "required": [ "name" ],
          "properties": {
            "name": {
              "type": "string"
            }
          }
        },
        "ValidationError": {
          "properties": {
            "type": {
              "type": "string",
              "enum": [ "ValidationError" ]
            },
            "details": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ValidationErrorDetail"
              }
            }
          }
        },
        "ValidationErrorDetail": {
          "properties": {
            "message": {
              "type": "string"
            },
            "path": {
              "type": "string"
            }
          }
        },
        "NotFoundError": {
          "properties": {
            "type": {
              "type": "string",
              "enum": [ "NotFoundError" ]
            },
            "details": {
              "type": "string",
              "enum": [ "User with id {id} not found" ]
            }
          }
        }
      }
    }
  }